Switched most long types to BigInteger, added BigInteger matrix operations. Possible some other small changes as well.

+++ Cryptosystem/src/adapters/MatrixAdapter.java
@@ -1,7 +1,9 @@
 package adapters;
 
+import java.math.BigInteger;
+
 public interface MatrixAdapter {
-	public long[][] add(long[][] m1, long[][] m2);
-	public long[][] multiply(long[][] m1, long[][] m2);
-	public long[][] transpose(long[][] m);
+	public BigInteger[][] add(BigInteger[][] m1, BigInteger[][] m2);
+	public BigInteger[][] multiply(BigInteger[][] m1, BigInteger[][] m2);
+	public BigInteger[][] transpose(BigInteger[][] m);
 }
