+++ Cryptosystem/src/homomorphicOperations/AddOps.java
@@ -1,5 +1,7 @@
 package homomorphicOperations;
 
+import java.math.BigInteger;
+
 import adapters.JBLASAdapter;
 import adapters.MatrixAdapter;
 
@@ -10,10 +12,10 @@ import adapters.MatrixAdapter;
  */
 
 public class AddOps {
-	public static long[][] add(long[][] c1, long[][] c2){
+	public static BigInteger[][] add(BigInteger[][] c1, BigInteger[][] c2){
 		//this should be a straightforward addition of the matricies, but we need to account for "overflow", which hasn't been done yet
 		MatrixAdapter ma = new JBLASAdapter();
-		long[][] c3 = ma.add(c1, c2);
+		BigInteger[][] c3 = ma.add(c1, c2);
 		return c3;
 	}
 }
